+++ src/main/java/org/aksw/iguana/cc/config/IguanaConfig.java
@@ -7,8 +7,8 @@ import org.aksw.iguana.commons.script.ScriptExecutor;
 import org.aksw.iguana.rp.controller.RPController;
 import org.aksw.iguana.rp.metrics.Metric;
 import org.aksw.iguana.rp.metrics.impl.*;
-import org.aksw.iguana.rp.storage.Storage;
-import org.aksw.iguana.rp.storage.impl.NTFileStorage;
+import org.aksw.iguana.cc.tasks.stresstest.storage.Storage;
+import org.aksw.iguana.cc.tasks.stresstest.storage.impl.NTFileStorage;
 import org.apache.commons.exec.ExecuteException;
 import org.apache.commons.lang3.SerializationUtils;
 import org.slf4j.Logger;
