+++ src/main/java/org/aksw/iguana/cc/tasks/stresstest/storage/impl/NTFileStorage.java
@@ -1,10 +1,10 @@
 /**
  * 
  */
-package org.aksw.iguana.rp.storage.impl;
+package org.aksw.iguana.cc.tasks.stresstest.storage.impl;
 
+import org.aksw.iguana.cc.tasks.stresstest.storage.TripleBasedStorage;
 import org.aksw.iguana.commons.annotation.Shorthand;
-import org.aksw.iguana.rp.storage.TripleBasedStorage;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFFormat;
 import org.slf4j.Logger;
