+++ src/main/java/org/aksw/iguana/cc/tasks/stresstest/storage/impl/RDFFileStorage.java
@@ -1,7 +1,7 @@
-package org.aksw.iguana.rp.storage.impl;
+package org.aksw.iguana.cc.tasks.stresstest.storage.impl;
 
+import org.aksw.iguana.cc.tasks.stresstest.storage.TripleBasedStorage;
 import org.aksw.iguana.commons.annotation.Shorthand;
-import org.aksw.iguana.rp.storage.TripleBasedStorage;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFLanguages;
