+++ src/test/java/org/aksw/iguana/cc/config/WorkflowTest.java
@@ -6,9 +6,9 @@ import org.aksw.iguana.commons.constants.COMMON;
 import org.aksw.iguana.rp.metrics.Metric;
 import org.aksw.iguana.rp.metrics.MetricManager;
 import org.aksw.iguana.rp.metrics.impl.*;
-import org.aksw.iguana.rp.storage.Storage;
-import org.aksw.iguana.rp.storage.StorageManager;
-import org.aksw.iguana.rp.storage.impl.NTFileStorage;
+import org.aksw.iguana.cc.tasks.stresstest.storage.Storage;
+import org.aksw.iguana.cc.tasks.stresstest.storage.StorageManager;
+import org.aksw.iguana.cc.tasks.stresstest.storage.impl.NTFileStorage;
 import org.apache.commons.io.FileUtils;
 import org.junit.After;
 import org.junit.Before;
