+++ src/test/resources/controller_test.properties
@@ -27,9 +27,9 @@ metric4.class=org.aksw.iguana.rp.metrics.impl.NoQPHMetric
 ##		Storages to use			##
 ##################################
 store1.class=org.aksw.iguana.rp.storage.imp.PropertiesSenderStorage
-store2.class=org.aksw.iguana.rp.storage.impl.NTFileStorage
+store2.class=org.aksw.iguana.cc.tasks.stresstest.storage.impl.NTFileStorage
 store2.constructorArgs=results_test.nt
-store3.class=org.aksw.iguana.rp.storage.impl.TriplestoreStorage
+store3.class=org.aksw.iguana.cc.tasks.stresstest.storage.impl.TriplestoreStorage
 store3.constructorArgs=http://localhost:3030/das/sparql,http://localhost:3030/das/update
 store4.class=org.aksw.iguana.rp.storage.impl.FileStorage
 store4.constructorArgs=result_storage
